app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
# Removed passlib[bcrypt] - using built-in hashlib instead
plaid-python==9.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2